убрал лишний модуль, перенёс константы
